+++ CHANGELOG.rst
@@ -30,6 +30,7 @@ Next version
   linting issues.
 - Removed custom HorizontalRule extension in favor of Tiptap's default
   implementation, which now includes proper insertion validation.
+- Exported Tiptap's ``Placeholder`` extension.
 
 
 0.16 (2025-07-11)